# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.addInstanceRequest:
    properties:
      fake_tls:
        description: pointer to distinguish unset from false
        type: boolean
      label:
        maxLength: 32
        type: string
      mask_host:
        type: string
      mask_port:
        maximum: 65535
        minimum: 1
        type: integer
      metrics_port:
        maximum: 65535
        minimum: 1
        type: integer
      port:
        maximum: 65535
        minimum: 1
        type: integer
      tags:
        description: JSON array
        type: string
      tcp_mss:
        maximum: 1460
        minimum: 1
        type: integer
      tcp_mss_enabled:
        type: boolean
      tls_domain:
        type: string
      tls_domains:
        description: JSON array
        type: string
      tls_fronting:
        type: boolean
    required:
    - port
    - tls_domain
    type: object
...
  handler.updateInstanceRequest:
    properties:
      enabled:
        type: boolean
      fake_tls:
        type: boolean
      label:
        maxLength: 32
        type: string
      mask_host:
        type: string
      mask_port:
        maximum: 65535
        minimum: 1
        type: integer
      metrics_port:
        maximum: 65535
        type: integer
      port:
        maximum: 65535
        minimum: 1
        type: integer
      tags:
        type: string
      tcp_mss:
        maximum: 1460
        minimum: 1
        type: integer
      tcp_mss_enabled:
        type: boolean
      tls_domain:
        type: string
      tls_domains:
        type: string
      tls_fronting:
        type: boolean
    type: object
...
host: localhost:8090
info:
  contact:
    name: PopuGate Support
    url: https://github.com/fussraider/PopuGate
  description: PopuGate is a Telegram MTProto proxy manager with a REST API.
  license:
    name: MIT
    url: https://github.com/fussraider/PopuGate/blob/main/LICENSE
  title: PopuGate API
  version: "1.0"
paths:
...
  /instances:
    get:
      description: Retrieve all proxy instances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List instances
      tags:
      - instances
    post:
      consumes:
      - application/json
      description: Create a new proxy instance with a port, optional metrics port,
        and label
      parameters:
      - description: Instance configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.addInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add an instance
      tags:
      - instances
  /instances/{id}:
    delete:
      description: Delete a proxy instance by ID. The last instance cannot be removed.
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove an instance
      tags:
      - instances
    put:
      consumes:
      - application/json
      description: Update an existing proxy instance configuration
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Instance updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateInstanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an instance
      tags:
      - instances
  /instances/{id}/logs:
    get:
      description: Retrieve logs for a specific proxy instance. Supports SSE streaming
        when follow=true.
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Number of log lines to return (default 100)
        in: query
        name: tail
        type: integer
      - description: Enable SSE streaming of log output
        in: query
        name: follow
        type: boolean
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Instance logs
      tags:
      - instances
  /instances/{id}/refresh-fronting:
    post:
      description: Re-download TLS fronting content for an instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Refresh TLS fronting content
      tags:
      - instances
  /instances/{id}/reload:
    post:
      description: Reload a specific proxy instance configuration
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reload an instance
      tags:
      - instances
  /instances/{id}/start:
    post:
      description: Start a specific proxy instance container
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start an instance
      tags:
      - instances
  /instances/{id}/status:
    get:
      description: Retrieve the current status of a specific proxy instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Instance status
      tags:
      - instances
  /instances/{id}/stop:
    post:
      description: Stop a specific proxy instance container
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Stop an instance
      tags:
      - instances
  /instances/check-port:
    get:
      description: Check if a TCP port is available for a new instance
      parameters:
      - description: Port number to check
        in: query
        name: port
        required: true
        type: integer
      - description: Instance ID to exclude from conflict check
        in: query
        name: exclude
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check port availability
      tags:
      - instances
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"